refactor: 대시보드 원본 데이터 커스텀 훅으로 분리
