+++ components/Edit/EditName.tsx
@@ -1,16 +1,13 @@
 import { useEffect, useState } from 'react'
 import { useRouter } from 'next/router'
-import { getDashBoardInfo, putDashBoard } from '@/service/dashboards'
+import { putDashBoard } from '@/service/dashboards'
 import { useAppDispatch } from '@/hooks/useApp'
 import { openToast } from '@/store/reducers/toastReducer'
+import useEditBoard from '@/hooks/useEditBoard'
 import { ShortButton } from '..'
 import ColorSelector from '../ColorSelector/ColorSelector'
 
 export default function EditName() {
-  const [dashboardBody, setDashBoardBody] = useState({
-    title: '',
-    color: '',
-  })
   const [editBoardBody, setEditBoardBody] = useState({
     title: '',
     color: '',
@@ -18,6 +15,7 @@ export default function EditName() {
   const router = useRouter()
   const { dashboardId } = router.query
   const dispatch = useAppDispatch()
+  const { dashboardBody } = useEditBoard(Number(dashboardId))
 
   const handleColor = (colorName: string) => {
     setEditBoardBody({
@@ -36,7 +34,6 @@ export default function EditName() {
     }
     dispatch(openToast('successEditBoard'))
     await putDashBoard(Number(dashboardId), editBoardBody)
-    setDashBoardBody({ ...editBoardBody })
     setEditBoardBody({
       ...editBoardBody,
       title: '',
@@ -44,15 +41,8 @@ export default function EditName() {
   }
 
   useEffect(() => {
-    const handleLoadDashBoard = async () => {
-      if (dashboardId) {
-        const { title, color } = await getDashBoardInfo(Number(dashboardId))
-        setDashBoardBody({ title, color })
-        setEditBoardBody({ title, color })
-      }
-    }
-    handleLoadDashBoard()
-  }, [dashboardId])
+    setEditBoardBody({ ...dashboardBody })
+  }, [dashboardBody])
 
   return (
     <div className="flex w-[62rem] flex-col gap-[3.4rem] rounded-[0.8rem] bg-var-white p-[2.8rem]">
